move helper func to util file
