+++ util.go
@@ -1,0 +1,22 @@
+package centrifuge
+
+func nextSeqGen(currentSeq, currentGen uint32) (uint32, uint32) {
+	var nextSeq uint32
+	nextGen := currentGen
+	if currentSeq == maxSeq {
+		nextSeq = 0
+		nextGen++
+	} else {
+		nextSeq = currentSeq + 1
+	}
+	return nextSeq, nextGen
+}
+
+func stringInSlice(a string, list []string) bool {
+	for _, b := range list {
+		if b == a {
+			return true
+		}
+	}
+	return false
+}
